test(Playwright): Move Core for Custom Matchers
It is moved from `aut`, which is not a fitting location for these
helpers to `expect`, assuming, that we will have some more
custom matchers.
